+++ src/main/java/com/yahoo/sketches/tuple/QuickSelectSketch.java
@@ -246,7 +246,7 @@ class QuickSelectSketch<S extends Summary> extends Sketch<S> {
     final byte[] bytes = new byte[sizeBytes];
     final Memory mem = new NativeMemory(bytes);
     int offset = 0;
-    mem.putByte(offset++, (byte) PREAMBLE_LONGS);
+    mem.putByte(offset++, PREAMBLE_LONGS);
     mem.putByte(offset++, serialVersionUID);
     mem.putByte(offset++, (byte) Family.TUPLE.getID());
     mem.putByte(offset++, (byte) SerializerDeserializer.SketchType.QuickSelectSketch.ordinal());
